+++ src/index.tsx
@@ -1,4 +1,5 @@
 import { render } from "preact";
 import App from "./App";
+import "./styles.css";
 
 render(<App/>, document.body);
